+++ test/test_helper.rb
@@ -60,7 +60,7 @@ class AppiumLibCoreTest
         platformName: :android,
         automationName: 'uiautomator2',
         app: 'test/functional/app/api.apk',
-        platformVersion: '7.1.1',
+        platformVersion: '8.1.0',
         deviceName: 'Android Emulator',
         appPackage: 'io.appium.android.apis',
         appActivity: 'io.appium.android.apis.ApiDemos',