+++ test/unit/android/device_test.rb
@@ -230,15 +230,6 @@ class AppiumLibCoreTest
         assert_requested(:post, "#{SESSION}/appium/performanceData/types", times: 1)
       end
 
-      def test_stop_recording_screen
-        stub_request(:post, "#{SESSION}/appium/stop_recording_screen")
-          .to_return(headers: HEADER, status: 200, body: { value: ['a'] }.to_json)
-
-        @driver.stop_recording_screen
-
-        assert_requested(:post, "#{SESSION}/appium/stop_recording_screen", times: 1)
-      end
-
       ## with args
 
       def test_available_contexts
@@ -476,13 +467,55 @@ class AppiumLibCoreTest
         assert_requested(:post, "#{SESSION}/appium/getPerformanceData", times: 1)
       end
 
-      def test_start_recording_screen
+      def test_start_recording_screen_default
         stub_request(:post, "#{SESSION}/appium/start_recording_screen")
+          .with(body: { options: { timeLimit: '180', bitRate: '4000000' } }.to_json)
           .to_return(headers: HEADER, status: 200, body: { value: ['a'] }.to_json)
 
         @driver.start_recording_screen
 
         assert_requested(:post, "#{SESSION}/appium/start_recording_screen", times: 1)
+      end
+
+      def test_start_recording_screen_custom
+        stub_request(:post, "#{SESSION}/appium/start_recording_screen")
+          .with(body: { options: { videoSize: '1280x1280', timeLimit: '60', bitRate: '5000000' } }.to_json)
+          .to_return(headers: HEADER, status: 200, body: { value: ['a'] }.to_json)
+
+        @driver.start_recording_screen video_size: '1280x1280', time_limit: '60', bit_rate: '5000000'
+
+        assert_requested(:post, "#{SESSION}/appium/start_recording_screen", times: 1)
+      end
+
+      def test_start_recording_screen_custom_false_force_restart
+        stub_request(:post, "#{SESSION}/appium/start_recording_screen")
+          .with(body: { options: { forceRestart: false, videoSize: '1280x1280', timeLimit: '60', bitRate: '5000000' } }.to_json)
+          .to_return(headers: HEADER, status: 200, body: { value: ['a'] }.to_json)
+
+        @driver.start_recording_screen video_size: '1280x1280', time_limit: '60', bit_rate: '5000000', force_restart: false
+
+        assert_requested(:post, "#{SESSION}/appium/start_recording_screen", times: 1)
+      end
+
+      def test_stop_recording_screen_default
+        stub_request(:post, "#{SESSION}/appium/stop_recording_screen")
+          .with(body: {}.to_json)
+          .to_return(headers: HEADER, status: 200, body: { value: ['a'] }.to_json)
+
+        @driver.stop_recording_screen
+
+        assert_requested(:post, "#{SESSION}/appium/stop_recording_screen", times: 1)
+      end
+
+      def test_stop_recording_screen_custom
+        stub_request(:post, "#{SESSION}/appium/stop_recording_screen")
+          .with(body: { options:
+                            { remotePath: 'https://example.com', user: 'user name', pass: 'pass', method: 'PUT' } }.to_json)
+          .to_return(headers: HEADER, status: 200, body: { value: ['a'] }.to_json)
+
+        @driver.stop_recording_screen(remote_path: 'https://example.com', user: 'user name', pass: 'pass')
+
+        assert_requested(:post, "#{SESSION}/appium/stop_recording_screen", times: 1)
       end
     end
   end